+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Added
+
+- Trophy Fish slot info - celeste
+
 ## 0.1.0-Alpha.10 - 2024-12-02
 
 ### Added